requirements.txt
aiogram~=3.1.0
Pillow~=10.0.1
reportlab~=4.0.5
python-docx~=0.8.11
pdf2docx~=0.5.6
selenium~=4.13.0
python-dotenv~=1.0.0
black~=23.9.1
